replace the native query for the overview list with JPQL query

+++ backend/src/main/java/xeus/com/vgdb/backend/persistence/dto/OverviewListItemDTO.java
@@ -12,46 +12,57 @@ public class OverviewListItemDTO {
     private Long gameId;
     private String title;
     private String gameDescription;
-    private Integer releasesCount;
+    private Long releasesCount;
     private GregorianCalendar firstReleaseDate;
 
     public Long getGameId() {
+
         return gameId;
     }
 
     public void setGameId(Long gameId) {
+
         this.gameId = gameId;
     }
 
     public String getTitle() {
+
         return title;
     }
 
     public void setTitle(String title) {
+
         this.title = title;
     }
 
     public String getGameDescription() {
+
         return gameDescription;
     }
 
     public void setGameDescription(String gameDescription) {
+
         this.gameDescription = gameDescription;
     }
 
-    public Integer getReleasesCount() {
+    public Long getReleasesCount() {
+
         return releasesCount;
     }
 
-    public void setReleasesCount(Integer releasesCount) {
+    public void setReleasesCount(Long releasesCount) {
+
         this.releasesCount = releasesCount;
     }
 
     public GregorianCalendar getFirstReleaseDate() {
+
         return firstReleaseDate;
     }
 
     public void setFirstReleaseDate(GregorianCalendar firstReleaseDate) {
+
         this.firstReleaseDate = firstReleaseDate;
     }
+
 }
